add render components and friend details component
