+++ friends/src/components/Friends/Header.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Header = () => {
+    return(
+        <div><h3>Address Book</h3></div>
+    );
+}
+
+export default Header;
